Add 'Map.save' method
